v.0.6.3: Added Framework target for OSX, Changed some ints to NSUInteger

+++ AFCache.m
@@ -549,7 +549,7 @@ static NSString *STORE_ARCHIVE_FILENAME = @ "urlcachestore";
 	return self;
 }
 
-- (unsigned)retainCount {
+- (NSUInteger)retainCount {
 	return UINT_MAX;  //denotes an object that cannot be released
 }
